track day

+++ firmware/src/App/Inc/controller_packet.h
@@ -14,11 +14,10 @@ typedef enum { KART_STATE_EBRAKE, KART_STATE_RC, KART_STATE_AUTO } Kart_State_T;
 
 typedef struct {
   Kart_State_T state;
-  int8_t throttle;
+  int16_t throttle;
   uint8_t steering;
 } Controller_Packet_T;
 
 Controller_Packet_T test = {.state = KART_STATE_EBRAKE};
 
 #endif /* APP_INC_CONTROLLER_PACKET_H_ */
-
